created Player.py, FallingObject.py modules, renamed the main file into "game.py" from "play.py"
